fixed signup with duplicate email address bug

+++ NMS-Cinema-UX/src/app/user-registration/user-registration.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { NgForm } from '@angular/forms';
 import { User } from '../models/user';
 import { UserService } from '../services/user.service';
 
@@ -10,6 +11,7 @@ import { UserService } from '../services/user.service';
 export class UserRegistrationComponent implements OnInit {
 
   user: User = new User();
+  success: boolean = true;
 
   constructor(private _userSvc: UserService) { }
 
@@ -17,19 +19,38 @@ export class UserRegistrationComponent implements OnInit {
   }
 
   register() {
-    console.log(this.user);
-    // TODO: Implement register function
+    // console.log(this.user);
 
-    this._userSvc.addUser(this.user).subscribe(
-      result=>{
-        sessionStorage.setItem("isUser", "true");
-        result.password = "";
-        sessionStorage.setItem("user", JSON.stringify(result));
+    // first see if another user exists with the submitted email address
+    let existingUser = null;
+    this._userSvc.getUserByEmail(this.user).subscribe(
+      result => {
+        existingUser = result;
+        if (existingUser.length == 0) {
+          // if no existing user, add the new one
+          this._userSvc.addUser(this.user).subscribe(
+            result => {
+              sessionStorage.setItem("isUser", "true");
+              result.password = "";
+              sessionStorage.setItem("user", JSON.stringify(result));
+            },
+            error => {
+              console.log("Error occured during registration." + error);
+              alert("An error during registration. Please try again later.");
+            });
+        }
+        else {
+          this.success = false;
+          this.user = new User();
+        }
       },
-      error=>{
-        console.log("Error occured during registration." + error);
-        alert("An error occured. Please try again later.");
+      error => {
+        console.log("an error occured while checking for an existing user account.");
       });
   }
 
+  resetRegistration(form:NgForm){
+    form.resetForm();
+    this.success=true;
+  }
 }
